+++ Tema02-ToDoList/main.js
@@ -1,0 +1,4 @@
+"use strict";
+exports.__esModule = true;
+var execute_1 = require("./execute");
+execute_1.execute();
