+++ Tema02-ToDoList/main.ts
@@ -1,0 +1,2 @@
+import { execute } from './execute';
+execute();
