+++ Tema02-ToDoList/todo.ts
@@ -1,0 +1,45 @@
+export class ToDo{
+    name: string;
+    endDate: Date;
+    finished: boolean;
+    description: string;
+
+    constructor(name: string, endDate: Date, description: string){
+        this.name = name;
+        this.endDate = endDate;
+        this.finished = false;
+        this.description = description;
+    }
+
+    //getters & setters
+    setName(name: string){
+        this.name = name;
+    }
+
+    setDate(date: Date){
+        this.endDate = date;
+    }
+
+    setDescription(description: string){
+        this.description = description;
+    }
+
+    getName(){
+        return this.name;
+    }
+
+    getDate(){
+        return this.endDate;
+    }
+
+    getDescrption(){
+        return this.description;
+    }
+
+    //other methods
+    finsh(){
+        this.finished = true;
+    }
+
+}
+
